requirements.txt
gradio==3.48.0
opencv-python==4.8.1.78
replicate==0.15.4
requests==2.31.0
python-dotenv==1.0.0